Confirmaçao js

+++ projeto/src/controller/cliente_bd/excluirCliente.php
@@ -3,15 +3,20 @@
 require_once "../../protect.php";
 require_once "../../conexao.php";
 
-$idCliente = isset($_POST['idcliente']) ? $_POST['idcliente'] : 0;
+if(isset($_SESSION['tipo']) && $_SESSION['tipo'] == "G"){
 
-if($idCliente > 0){ 
-    $sql_code = "DELETE FROM cliente WHERE idcliente = '$idCliente' ";
-    $sql_query = $conexao->query($sql_code);
+    $idCliente = isset($_GET['id']) ? $_GET['id'] : 0;
 
-    header("Location: ../../../clientes.php");
+    if($idCliente > 0){ 
+        $sql_code = "DELETE FROM cliente WHERE idcliente = '$idCliente' ";
+        $sql_query = $conexao->query($sql_code);
 
+        header("Location: ../../../clientes.php");
 
+    } else {
+        header("location: ../../../index.php");
+    }
 } else {
-    header("location: ../../../index.php");
+    header("Location: ../../../nao_permitido.php");
 }
+
